+++ due/src/utils/splitter.rs
@@ -5,14 +5,14 @@ use tokio::sync::mpsc::{unbounded_channel, UnboundedReceiver, UnboundedSender};
 use crate::packets::packet::Packet;
 
 pub struct Splitter {
-    element_id: u32,
+    element_id: usize,
     pub sender_1: UnboundedSender<Packet>,
     pub sender_2: UnboundedSender<Packet>,
     pub receiver: UnboundedReceiver<Packet>,
 }
 
 impl Splitter {
-    pub fn new(element_id: u32) -> Splitter {
+    pub fn new(element_id: usize) -> Splitter {
         Splitter {
             element_id,
             sender_1: unbounded_channel().0,
